Update awk fields to tabs
The jq call is already returning tab-delimited data, if awk printed a substr of $6 and that happens to be a digit it would print the entire $0 again and throw off proper printing. Now we just print $6 to include the entire file/folder name, even if it has spaces in it.